BUILD: Fix PSP compiler failing during C++ version tests
The blessed way to run the compiler is to use some custom GNU Make
file from PSPSDK, which we do not use, which sets the appropriate
linker flags. Running the compiler stand-alone just fails at link
time because crt0.o references libraries which are not included
by default.

The main compiler test was changed to not link
in 513cdb44a6bf866a893275538a19bfee8c7f465d, apparently because the
PS2SDK did the same thing, and now that we have more tests (some
of which do require linking since we test for unresolvable
symbols), it is really necessary to get the flags set up early.