GLK: FROTZ: Use the Noto runic font for runes
We already bundle other Noto fonts in the font.dat file, so using
further Noto fonts makes sense to me. Also, map upper-case letters
to lower-case runes since there are versions of Beyond Zork that
uses that. (The version I played many years ago did, and it looked
very strange. The version I tested with now did not, probably for
that very reason. So that part is untested for now.)