DEVTOOLS: Multilingual support for quoteSpreadsheetCreator
Versions supported are English, German, French, Italian and Spanish

More work would be required for Russian versions to be supported by the spreadsheet creator.
I don't have access to the Russian version that ScummVM currently supports/in the detection table.
Current implementation could be improved (to do away with all the if clauses for the glyph "exceptions")
but this is not high priority.