DEVTOOLS: Handle empty AUD file in Blade Runner
French localization has an empty AUD file. Only the header is present and seems corrupted

This is DCLER_0030_00041ECE in TLK0A. It is an unused audio file ("boop" in ENG localization), but skipping it would cause issues with the translation excel, ie. one row missing. The fix is to create a placeholder empty wav file in order to not ignore this empty quote. The bug affected the spreasheet creator tool.