KYRA: (EOB) - fix unused string
This fixes the "%s can not use this item" string. AFAIK this message never appears. It would require to place an item into a character's hand which he/she can't use and right-click use it. However, when you place such an unusable item in a character's hand the slot will be greyed out/disabled immediately and can no longer be clicked.